Renamed routing files to better align with fastAPI's nomenclature.
