+++ backend/app/main.py
@@ -2,8 +2,9 @@
 from fastapi import FastAPI, HTTPException
 from starlette.concurrency import run_in_threadpool
 
-from backend.app.routes.emails_routes import router as emails_router
-from backend.app.routes.summaries_routes import router as summaries_router
+from backend.app.routers.emails_router import router as emails_router
+from backend.app.routers.summaries_router import router as summaries_router
+from backend.app.routers.auth_router import router as auth_router
 
 from backend.app.models.email_model import Email
 from backend.app.models.summary_model import EmailSummary
